Read ASDO and TDO pins one clock cycle after writes
Previously the CPLD logic was reading them at the same time as writing
new pin states. This seemed to work okay inside of Windows, but it
didn't work in Linux. Read the pins one CLK cycle later. It works
better.

This kind of makes sense. If we are updating TCK to be low, that will
cause the connected JTAG device to update its output TDO state. Wait a
single CLK cycle so that TDO is stable before reading it back.

My guess is Altera's JTAG software on Windows waits until the next
transaction to read back the TDO state, but on Linux it expects the
readback to be valid in the same transaction. I could be wrong, but this
change definitely seems to improve the situation.